fix(gh): bound error text for settings, labels, and licence calls
boundedActionsHTTPError truncated and stripped control characters from
GitHub API error text for Actions endpoints only, so settings, labels,
and licence error bodies reached the terminal unbounded.

- Move the helper to internal/gh/errors.go as boundedHTTPError, unchanged in
  limit (256 characters, 3 details) and behaviour
- Wrap every settings, labels, and licence API call, alongside the existing
  Actions call sites
- Preserve the *api.HTTPError type so status-code logic, such as treating
  404 as disabled, is unchanged
- Add TestFetchLicenceBoundsAPIErrorBody to prove a control-laden 500-character
  body is bounded
