+++ internal/gh/licence.go
@@ -14,7 +14,7 @@ type licenceResponse struct {
 // FetchLicence returns the licence body text from GET /licenses/{id}.
 func FetchLicence(client *api.RESTClient, id string) (string, error) {
 	var resp licenceResponse
-	if err := client.Get(fmt.Sprintf("licenses/%s", id), &resp); err != nil {
+	if err := boundedHTTPError(client.Get(fmt.Sprintf("licenses/%s", id), &resp)); err != nil {
 		return "", fmt.Errorf("fetching licence %q: %w", id, err)
 	}
 	return resp.Body, nil
